Move background script to dedicated folder
