+++ src/js/background/index.ts
@@ -1,0 +1,8 @@
+import "../../img/icon-128.png";
+import "../../img/icon-34.png";
+
+import openOptionsPageOnInstall from "./open-options-page-on-install";
+import allowTaggerInIframe from "./allow-tagger-in-iframe";
+
+openOptionsPageOnInstall();
+allowTaggerInIframe();
