+++ src/js/background/open-options-page-on-install.js
@@ -1,0 +1,9 @@
+import { onInstalled, openOptionsPage } from "Browser/runtime";
+
+export default function () {
+  onInstalled().addListener(function (details) {
+    if (details.reason === "install") {
+      openOptionsPage();
+    }
+  });
+}
